refactor: Update imports to use '@/lib' instead of '@/lib/utils' across multiple components; adjust Vite config aliasing for improved clarity
