+++ react-app/src/components/ui/spinner.tsx
@@ -4,7 +4,7 @@ import {
   LoaderPinwheelIcon,
   type LucideProps,
 } from 'lucide-react';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib';
 
 type SpinnerVariantProps = Omit<SpinnerProps, 'variant'>;
