+++ react-app/vite.config.ts
@@ -7,8 +7,12 @@ import tailwindcss from '@tailwindcss/vite';
 export default defineConfig({
   plugins: [react(), tailwindcss()],
   resolve: {
-    alias: {
-      '@': path.resolve(__dirname, './src'),
-    },
+    alias: [
+      {
+        find: '@',
+        replacement: path.resolve(__dirname, './src'),
+      },
+    ],
+    extensions: ['.ts', '.tsx', '.js', '.jsx', '.json'],
   },
 });
